# be/cmd/server/docs/swagger.yaml
definitions:
...
  dto.CheckPasswordReset:
    properties:
      email:
        type: string
      redirectUrl:
        type: string
    required:
    - email
    - redirectUrl
    type: object
...
  dto.RefreshRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  dto.UpdateInformationUserRequest:
    properties:
      firstName:
        type: string
      lastName:
        type: string
    required:
    - firstName
    - lastName
    type: object
  dto.UpdateRoleUserRequest:
    properties:
      roleTitle:
        type: string
    required:
    - roleTitle
    type: object
  dto.UserLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.UserRegisterRequest:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
      redirectUrl:
        type: string
    required:
    - email
    - firstName
    - lastName
    - password
    - redirectUrl
    type: object
  dto.UserRequestResetPassword:
    properties:
      newPassword:
        type: string
      token:
        type: string
    required:
    - newPassword
    - token
    type: object
info:
  contact: {}
paths:
...
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Login
      parameters:
      - description: Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UserLoginRequest'
      produces:
      - application/json
      responses: {}
      summary: Login
      tags:
      - Auth
...
  /api/auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh Token
      parameters:
      - description: Data
        in: body
        name: refresh_token
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshRequest'
      produces:
      - application/json
      responses: {}
      summary: Refresh Token
      tags:
      - Auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Register user
      parameters:
      - description: Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UserRegisterRequest'
      produces:
      - application/json
      responses: {}
      summary: Register user
      tags:
      - Auth
...
  /api/user/{email}:
    delete:
      consumes:
      - application/json
      description: Delete user via email
      parameters:
      - description: email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Delete user
      tags:
      - Users
  /api/user/forget-password:
    post:
      consumes:
      - application/json
      description: Email reset password
      parameters:
      - description: Data
        in: body
        name: Email_Reset_Password
        required: true
        schema:
          $ref: '#/definitions/dto.CheckPasswordReset'
      produces:
      - application/json
      responses: {}
      summary: Email reset password
      tags:
      - Users
  /api/user/information:
    patch:
      consumes:
      - application/json
      description: Update Information
      parameters:
      - description: Data
        in: body
        name: Information
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateInformationUserRequest'
      produces:
      - application/json
      responses: {}
      summary: Update Information
      tags:
      - Users
  /api/user/password-reset:
    patch:
      consumes:
      - application/json
      description: reset password
      parameters:
      - description: Data
        in: body
        name: Password-reset
        required: true
        schema:
          $ref: '#/definitions/dto.UserRequestResetPassword'
      produces:
      - application/json
      responses: {}
      summary: Password-reset
      tags:
      - Users
  /api/user/role:
    patch:
      consumes:
      - application/json
      description: Update Role
      parameters:
      - description: Data
        in: body
        name: Role
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateRoleUserRequest'
      produces:
      - application/json
      responses: {}
      summary: Update Role
      tags:
      - roles
...
swagger: "2.0"